"Added models package and updated GetMenu function to find and return a single menu by ID"
This commit updates the `GetMenu` function to find and return a single menu by its ID. It adds a new import statement for the `models` package and uses it to define a `menu` variable of type `models.Menu`. The function then finds the menu with the specified ID using the `FindOne` method and returns it in the response. If there is an error while finding the menu, the function responds with an HTTP 500 status code and an error message. Otherwise, it responds with an HTTP 200 status code and the menu data. The `context.WithTimeout` function is used to set a timeout for the database operation.